+++ hgext/bzpost/__init__.py
@@ -169,7 +169,7 @@ def wrappedpushbookmark(orig, pushop):
         comment = '\n'.join(lines)
 
         ui.write(_('recording push in bug %s\n') % bugnumber)
-        bug.add_comment_working(comment)
+        bug.add_comment(comment)
 
     return result