+++ pylib/Bugsy/bugsy/__init__.py
@@ -1,3 +1,3 @@
-from bug import Bug, BugException
+from bug import Bug, BugException, Comment
 from bugsy import Bugsy, BugsyException, LoginException
 from search import Search